+++ inst/app/global.R
@@ -64,12 +64,16 @@ if (!"package:radiant.data" %in% search())
 ## running local or on a server
 if (Sys.getenv('SHINY_PORT') == "") {
   options(radiant.local = TRUE)
+  options(radiant.report = getOption("radiant.report", default = TRUE)) 
   ## no limit to filesize locally
-  options(shiny.maxRequestSize = -1)
+  # options(shiny.maxRequestSize = -1)
+  options(shiny.maxRequestSize = getOption("radiant.maxRequestSize", default = -1))
 } else {
   options(radiant.local = FALSE)
+  options(radiant.report = getOption("radiant.report", default = FALSE)) 
   ## limit upload filesize on server (10MB)
-  options(shiny.maxRequestSize = 10 * 1024^2)
+  # options(shiny.maxRequestSize = 10 * 1024^2)
+  options(shiny.maxRequestSize = getOption("radiant.maxRequestSize", default = 10 * 1024^2))
 }
 
 ## encoding
@@ -246,4 +250,16 @@ if (length(tmp) > 0)
 options(radiant.versions = paste(radiant.versions, collapse = ", "))
 rm(tmp, radiant.versions)
 
+navbar_proj <- function(navbar) {
+  ## getting project information
+  proj <- rstudioapi::getActiveProject()
+  proj <- if (is.null(proj)) "Project: (None)" else paste0("Project: ", basename(proj))
+  if (!isTRUE(getOption("radiant.local"))) proj <- "Remote session" 
+  proj <- tags$span(class = "nav navbar-brand navbar-right", proj)
 
+  ## based on: https://stackoverflow.com/a/40755608/1974918
+  navbar[[3]][[1]]$children[[1]]$children[[2]] <- htmltools::tagAppendChild(
+    navbar[[3]][[1]]$children[[1]]$children[[2]], proj)
+
+  navbar
+}