+++ inst/app/ui.R
@@ -1,5 +1,4 @@
 ## ui for data menu in radiant
-do.call(navbarPage,
-  c("Radiant", getOption("radiant.nav_ui"), getOption("radiant.shared_ui"),
-    help_menu("help_data_ui"))
-)
+navbar_proj(do.call(navbarPage, c("Radiant", getOption("radiant.nav_ui"),
+  getOption("radiant.shared_ui"), help_menu("help_data_ui"))
+))